moved code from useEns composable to component (composable won't be reused anyway)
